requirements.txt
anyio==3.5.0
asgiref==3.5.0
black==21.10b0
CacheControl==0.12.9
cachetools==4.2.4
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
et-xmlfile==1.1.0
fastapi==0.73.0
fastpip==0.12.1
firebase-admin==5.1.0
Flask==2.0.2
fpdf==1.7.2
ghostscript==0.7
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
h11==0.13.0
httplib2==0.20.2
httptools==0.3.0
idna==3.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
numpy==1.21.4
opencv-python==4.5.4.58
openpyxl==3.0.9
packaging==21.2
pandas==1.3.4
pathspec==0.9.0
pdf2image==1.16.0
pdfminer.six==20211012
Pillow==8.2.0
platformdirs==2.4.0
proto-plus==1.19.7
protobuf==3.19.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.9.0
pyparsing==3.0.4
PyPDF2==1.26.0
pytesseract==0.3.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tabulate==0.8.9
tomli==1.2.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.17.3
watchgod==0.7
websockets==10.1
Werkzeug==2.0.2
